Consequencify active-state blinking
