+++ Assets/Scenes/ProdScenes/Level 1.unity
@@ -84157,8 +84157,9 @@ GameObject:
   m_Component:
   - component: {fileID: 1831325032}
   - component: {fileID: 1831325033}
+  - component: {fileID: 1831325034}
   m_Layer: 0
-  m_Name: Blinker
+  m_Name: BlinkyLasers
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -84195,6 +84196,21 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   minSecondsBetweenSwitching: 2
   maxSecondsBetweenSwitching: 4
+--- !u!114 &1831325034
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1831325031}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 16b56fc371fbf8c4d80e0245173696d3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  destroy: 0
+  consequenceObject: {fileID: 0}
+  findConsequencesInChildren: 0
 --- !u!1001 &1833017966
 PrefabInstance:
   m_ObjectHideFlags: 0
